fix: Move HTTPX client cleanup from agent to browser close to avoid breaking multi-agent task
